commented broken scorecard stuff

+++ db/migrate/20161024173006_create_user_scorecards.rb
@@ -2,10 +2,10 @@ class CreateUserScorecards < ActiveRecord::Migration[5.0]
   def change
     create_table :user_scorecards do |t|
       t.integer :volunteer_user_id
-      t.integer :skill_proficiency
-      t.integer :attitude
-      t.integer :enthusiasm
-      t.integer :reliability
+      t.float :skill_proficiency
+      t.float :attitude
+      t.float :enthusiasm
+      t.float :reliability
 
       t.timestamps
     end
